+++ src/Products/PloneMeeting/tests/testMeetingItem.py
@@ -8181,6 +8181,13 @@ class testMeetingItem(PloneMeetingTestCase):
         vendors_item = self.create('MeetingItem')
         self.assertEqual(vendors_item.Vocabulary('committees')[0].keys(),
                          [NO_COMMITTEE, 'committee_2', u'committee_for_item'])
+        # when committee no accessible but stored on context it is part of the vocabulary
+        self.assertFalse(cfg_committees[0]['row_id'] in
+                         vendors_item.Vocabulary('committees')[0].keys())
+        vendors_item.setCommittees((cfg_committees[0]['row_id'], ))
+        import ipdb; ipdb.set_trace()
+        self.assertTrue(cfg_committees[0]['row_id'] in
+                        vendors_item.Vocabulary('committees')[0].keys())
 
     def test_pm_CommitteesItemOnly(self):
         """It is possible to display a committee only on the item and not on the meeting,